teraranger: move to PX4Rangefinder and cleanup
