+++ src/drivers/distance_sensor/teraranger/CMakeLists.txt
@@ -1,6 +1,6 @@
 ############################################################################
 #
-#   Copyright (c) 2017 PX4 Development Team. All rights reserved.
+#   Copyright (c) 2017-2020 PX4 Development Team. All rights reserved.
 #
 # Redistribution and use in source and binary forms, with or without
 # modification, are permitted provided that the following conditions
@@ -31,12 +31,13 @@
 #
 ############################################################################
 px4_add_module(
-	MODULE drivers__teraranger
+	MODULE drivers__distance_sensor__teraranger
 	MAIN teraranger
-	COMPILE_FLAGS
-		-Wno-cast-align # TODO: fix and enable
 	SRCS
-		teraranger.cpp
+		teraranger_main.cpp
+		TERARANGER.cpp
+		TERARANGER.hpp
 	DEPENDS
+		drivers_rangefinder
+		px4_work_queue
 	)
-
